refactor: gorup analysis and market data tools mounting mcp
