+++ src/tools/analysis.py
@@ -1,8 +1,68 @@
 import yfinance as yf
 from typing import Dict
+from fastmcp import FastMCP
+
+analysis_mcp = FastMCP("Analysis")
+
+@analysis_mcp.tool
+def fetch_fundamentals(ticker_id: str) -> Dict[str, float]:
+    """
+    Fetches fundamental data for a given ticker.
+
+    Parameters:
+    ticker_id (str): e.g. 'AAPL' or 'GOOG'
+
+    Returns:
+    dict: Dictionary with Market Cap, PE Ratio, EPS, Revenue, and Earnings.
+    """
+    try:
+        ticker = yf.Ticker(ticker_id)
+        info = ticker.info
+
+        fundamentals_data = {
+            "market_cap": info.get("marketCap"),
+            "pe_ratio": info.get("trailingPE"),
+            "eps": info.get("trailingEps"),
+            "revenue": info.get("totalRevenue"),
+            "earnings": info.get("grossProfits")
+        }
+
+        return fundamentals_data
+
+    except Exception as e:
+        return {"error": str(e)}
+
+@analysis_mcp.tool
+def fetch_earnings_dates(ticker_id: str) -> dict:
+    """
+    Fetches upcoming earnings announcement dates for a given ticker.
+
+    Parameters:
+    ticker_id (str): e.g. 'AAPL' or 'GOOG'
+
+    Returns:
+    dict: Dictionary with earnings dates and related information.
+    """
+    try:
+        ticker = yf.Ticker(ticker_id)
+        calendar = ticker.earnings_dates
+
+        if calendar is None or calendar.empty:
+            calendar = ticker.calendar
+
+        if calendar is not None and not calendar.empty:
+            earnings_dict = calendar.to_dict()
+            return earnings_dict
+        else:
+            return {"message": "No earnings calendar data available."}
+
+    except Exception as e:
+        return {"error": str(e)}
 
 
-def get_technical_indicators(ticker, start, end) -> Dict[str, float]:
+
+@analysis_mcp.tool
+def fetch_technical_indicators(ticker: str, start: str, end: str) -> Dict[str, float]:
     """
     Fetches technical indicators for a ticker using yfinance and pandas.
 
@@ -76,5 +136,4 @@ def get_technical_indicators(ticker, start, end) -> Dict[str, float]:
         return indicators.dropna().to_dict()
 
     except Exception as e:
-        print(f"Error fetching indicators for {ticker}: {e}")
-        return None
+        return {"error": str(e)}
